ADM-365 refactor pipelineMetricsSettings to metricsSlice
