+++ frontend/src/components/Metrics/MetricsStep/DeploymentFrequencySettings/SingleSelection/index.tsx
@@ -2,7 +2,7 @@ import { FormHelperText, InputLabel, ListItemText, MenuItem, Select, SelectChang
 import React, { useState } from 'react'
 import { FormControlWrapper } from './style'
 import { useAppDispatch } from '@src/hooks'
-import { updateDeploymentFrequencySettings } from '@src/context/pipelineMetricsSettings/pipelineMetricsSettingsSlice'
+import { updateDeploymentFrequencySettings } from '@src/context/Metrics/metricsSlice'
 
 interface Props {
   options: string[]
